proxy: move metrics::prometheus module to root metrics module
