+++ proxy/src/telemetry/metrics/labels.rs
